Handled excess positional arguments and invalid attribute names
This update handles cases where the number of positional arguments
passed to the `update()` method is more than the number of available
attributes. It also stops the addition of new attributes that do not
exist in the original `__dict__`. The idea is to update the already
existing values, so if it does not exist, it shouldn't be created.