feat: Fetch and display StarWars character name on the webpage
This commit adds functionality to fetch the StarWars character name from the API endpoint 'https://swapi-api.alx-tools.com/api/people/5/?format=json' and display it on the webpage. The character name is appended to the element with the id 'character'. This enhancement allows users to see the name of the StarWars character on the webpage.