feat: Add functionality to add and remove items from a list
This commit adds event listeners to the `div#add_item` and `div#remove_item` elements. When the `div#add_item` element is clicked, an `<li>` element with the text "Item" is appended to the `ul.my_list` element. When the `div#remove_item` element is clicked, the last `<li>` element in the `ul.my_list` is removed. This enhancement allows users to easily add and remove items from the list by clicking on the designated elements.